fixed the weird issue with the polyline drawing
